refactor(cloudflare): extract shared HTTP client builders for D1, Flagship, Browser
Same convention as KV/R2/Tunnel/AI/Vectorize: move client construction into
unexported {Cap}HttpClient.ts scaffolding parameterized by an injectable
auth ({ authorize, accountId }), so a future token-scoped *Http layer reuses
the builder instead of duplicating it. The *Local layers are now just
layer + auth + deferred id accessor.
